requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
datasets==2.16.1
dill==0.3.7
einops==0.7.0
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.10.0
huggingface-hub==0.20.2
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.2.1
numpy==1.26.3
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pyarrow==14.0.2
pyarrow-hotfix==0.6
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sympy==1.12
tokenizers==0.15.0
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
xxhash==3.4.1
yarl==1.9.4
